+++ src/realm/link_view.cpp
@@ -309,7 +309,7 @@ TableView LinkView::get_sorted_view(std::vector<size_t> column_indexes, std::vec
     TableView v(m_origin_column.get_target_table()); // sets m_table
     v.m_last_seen_version = m_origin_table->m_version;
     // sets m_linkview_source to indicate that this TableView was generated from a LinkView
-    v.m_linkview_source = ConstLinkViewRef(this);
+    v.m_linkview_source = shared_from_this();
     if (m_row_indexes.is_attached()) {
         for (size_t t = 0; t < m_row_indexes.size(); t++) // todo, simpler way?
             v.m_row_indexes.add(get(t).get_index());